Separated javascript into proper static media. Upgraded tag-it.
