+++ taggit_autocomplete/settings.py
@@ -4,4 +4,5 @@ from django.conf import settings
 CSS = getattr(settings, 'TAGGIT_AUTOCOMPLETE_CSS', ())
 CSS = CSS + ('taggit_autocomplete/css/jquery.tagit.css', )
 JS = getattr(settings, 'TAGGIT_AUTOCOMPLETE_JS', ())
-JS = JS + ('taggit_autocomplete/js/tag-it.js', )
+JS = JS + ('taggit_autocomplete/js/tag-it.min.js', 'taggit_autocomplete/js/django-taggit-ac.js',)
+LIMIT = getattr(settings, 'TAGGIT_AUTOCOMPLETE_LIMIT', 10)
