+++ taggit_autocomplete/views.py
@@ -5,6 +5,7 @@ except ImportError:
     import json as simplejson  # noqa
 
 from taggit.models import Tag
+from taggit_autocomplete import settings
 
 
 def list_tags(request):
@@ -14,12 +15,7 @@ def list_tags(request):
     if request.method != 'GET' or not query:
         return HttpResponseBadRequest()
 
-    try:
-        tags = Tag.objects.filter(
-            name__istartswith=query).values_list('name', flat=True)
-    except:  # Naked except, we may not have access to hvad
-        tags = Tag.objects.language().filter(
-            name__istartswith=query).values_list('name', flat=True)
+    tags = Tag.objects.filter(name__istartswith=query).values_list('name', flat=True)[:settings.LIMIT]
 
     tags = list(tags)
     return HttpResponse(simplejson.dumps(tags), content_type='text/javascript')
