+++ taggit_autocomplete/widgets.py
@@ -10,36 +10,12 @@ class TagAutocomplete(forms.TextInput):
     input_type = 'text'
 
     def render(self, name, value, attrs=None):
-        list_view = reverse('taggit_autocomplete-list')
         if value is not None and not isinstance(value, basestring):
             value = edit_string_for_tags(
                     [o.tag for o in value.select_related("tag")])
-        html = super(TagAutocomplete, self).render(name, value, attrs)
-
-        # Activate tag-it in this field
-        js = u"""
-            <script type="text/javascript">
-                (function($) {
-                    $(document).ready(function() {
-                        $("#%(id)s").tagit({
-                            caseSensitive: false,
-                            tagSource: function(search, showChoices) {
-                                options = this;
-                                $.getJSON("%(source)s", {
-                                    q: search.term.toLowerCase()
-                                }, function(data) {
-                                    showChoices(options._subtractArray(data, options.assignedTags()));
-                                });
-                            }
-                        });
-                    });
-                })(jQuery);
-            </script>
-            """ % ({
-                'id': attrs['id'],
-                'source': list_view
-            })
-        return mark_safe("\n".join([html, js]))
+        attrs['class'] = (attrs.get('class', '') + ' django-taggit-ac').strip()
+        attrs['data-src'] = reverse('taggit_autocomplete-list')
+        return super(TagAutocomplete, self).render(name, value, attrs)
 
     class Media:
         css = {
